[FIX] virt: Difference in port callback argument width between C# and C

+++ src/Emulator/Cores/virt/include/callbacks.h
@@ -27,13 +27,13 @@
 void kvm_log(int level, char *message);
 void kvm_abort(char *message);
 
-uint64_t kvm_io_port_read_byte(uint64_t address);
-uint64_t kvm_io_port_read_word(uint64_t address);
-uint64_t kvm_io_port_read_double_word(uint64_t address);
+uint32_t kvm_io_port_read_byte(uint32_t address);
+uint32_t kvm_io_port_read_word(uint32_t address);
+uint32_t kvm_io_port_read_double_word(uint32_t address);
 
-void kvm_io_port_write_byte(uint64_t address, uint64_t value);
-void kvm_io_port_write_word(uint64_t address, uint64_t value);
-void kvm_io_port_write_double_word(uint64_t address, uint64_t value);
+void kvm_io_port_write_byte(uint32_t address, uint32_t value);
+void kvm_io_port_write_word(uint32_t address, uint32_t value);
+void kvm_io_port_write_double_word(uint32_t address, uint32_t value);
 
 uint64_t kvm_sysbus_read_byte(uint64_t address);
 uint64_t kvm_sysbus_read_word(uint64_t address);
